fix(T11455): host background-jobs in @cleocode/core to fix dual drizzle-instance tsc -b error
CI Type Check (tsc -b, full project-reference graph) caught a dual drizzle-orm
instance type mismatch (TS2345/TS2769 on every query builder in the relocated
background-jobs.ts): adding drizzle-orm as a regular dependency of
@cleocode/runtime made pnpm resolve a different peer-hashed drizzle instance
than @cleocode/core's, so the runtime-constructed SQL<unknown> was nominally
incompatible with core's .where() overloads.

Architectural fix: background-jobs.ts is Drizzle DB-access logic, so it now
lives in @cleocode/core/store (the package that owns the consolidated schema +
the SINGLE drizzle instance + the DB handle types). Inside core its imports are
local (./sqlite.js, ./tasks-schema.js), so the query builders type-check
against one instance. @cleocode/runtime/gateway/background-jobs.ts is now a thin
re-export of @cleocode/core/store/background-jobs, the job-manager accessor
imports the type from core, and drizzle-orm is REMOVED from @cleocode/runtime's
dependencies (pnpm-lock returns to pristine). Runtime still has no
@cleocode/cleo dependency (AC3) and the decouple stays complete.

Validated with `pnpm run typecheck` (tsc -b): exit 0, ZERO errors. Also: full
build, cleo check arch (5/5), Gate 11, transport-inventory golden (11/11),
biome, frozen-lockfile — all green.
