+++ packages/runtime/package.json
@@ -25,8 +25,7 @@
   },
   "dependencies": {
     "@cleocode/contracts": "workspace:*",
-    "@cleocode/core": "workspace:*",
-    "drizzle-orm": "1.0.0-rc.3"
+    "@cleocode/core": "workspace:*"
   },
   "devDependencies": {
     "tsup": "^8.0.0",
